% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancytable.R
\name{fancytable}
\alias{fancytable}
\title{A fancy table with proportions}
\usage{
fancytable(tab, orientation = "h", dec = 2)
}
\arguments{
\item{tab}{table object input}

\item{orientation}{orientation character: "v", "h" (default) or "t"}

\item{dec}{Number. How many decimals to output (default:2)}
}
\value{

}
\description{
A fancy table with proportions
}
