% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum.R
\name{lipsum}
\alias{lipsum}
\title{Insert filler text in document}
\usage{
lipsum(n = 1)
}
\arguments{
\item{n}{Number of paragraphs}
}
\value{

}
\description{
Insert filler text in document
}
\examples{
lipsum(3)
lipsum(10)
}
