% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancytable.R
\name{ornt}
\alias{ornt}
\title{Orientation aux function}
\usage{
ornt(tab, ori)
}
\arguments{
\item{tab}{table input}

\item{ori}{orientation character: "v" or "h" or "t"}
}
\value{

}
\description{
Orientation aux function
}
