% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelaum.R
\name{tabelaum}
\alias{tabelaum}
\title{Create big table for reports}
\usage{
tabelaum(dep_var, list_cont = NULL, list_cat = NULL, paired = FALSE, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Create big table for reports
}
\examples{
tabelaum(dep_var = "vs",
list_cont=c("mpg","cyl"),
list_cat = "am",
data = mtcars)

tabelaum(dep_var = "Species",
         list_cont=c(1:4),
         data=iris)

}
